refac(package): split package into files, load default auth file in package
