+++ image/errors.py
@@ -1,0 +1,7 @@
+"""
+ContainerImageError exception
+
+A generic Exception extension for errors thrown by containerimage-py
+"""
+class ContainerImageError(Exception):
+    pass
