+++ image/mediatypes.py
@@ -1,0 +1,7 @@
+# Container image media types
+DOCKER_V2S2_MEDIA_TYPE = "application/vnd.docker.distribution.manifest.v2+json"
+DOCKER_V2S1_SIGNED_MEDIA_TYPE = "application/vnd.docker.distribution.manifest.v1+prettyjws"
+DOCKER_V2S1_MEDIA_TYPE = "application/vnd.docker.distribution.manifest.v1+json"
+DOCKER_V2S2_LIST_MEDIA_TYPE = "application/vnd.docker.distribution.manifest.list.v2+json"
+OCI_MANIFEST_MEDIA_TYPE = "application/vnd.oci.image.manifest.v1+json"
+OCI_INDEX_MEDIA_TYPE = "application/vnd.oci.image.index.v1+json"
